requirements.txt
anyio==3.6.2
biopython==1.79
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
fastapi==0.86.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
seaborn==0.12.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.4
